+++ apps/nowcasting-app/pages/index.tsx
@@ -4,8 +4,10 @@ import Layout from "../components/layout";
 import { PvLatestMap } from "../components/map";
 import SideLayout from "../components/side-layout";
 import PvRemixChart from "../components/charts/pv-remix-chart";
+import useAndUpdateSelectedTime from "../components/hooks/use-and-update-selected-time";
 
 export default function Home() {
+  useAndUpdateSelectedTime();
   return (
     <Layout>
       <div className="h-full flex">
